move and rename post item
