+++ src/app/components/post/view/view.vue
@@ -13,35 +13,7 @@
 					<div class="post-view">
 						<div v-if="post.hasMedia" class="full-bleed-xs">
 							<div v-for="item of post.media" :key="item.id">
-								<app-responsive-dimensions
-									class="-media-item"
-									:ratio="item.width / item.height"
-									:max-width="item.width"
-								>
-									<app-event-item-media-tags :gif="item.is_animated" />
-									<app-media-item-backdrop
-										class="-backdrop"
-										:media-item="item"
-										:radius="Screen.isXs ? null : 'lg'"
-									>
-										<app-img-responsive
-											class="-img"
-											v-if="!item.is_animated"
-											:src="item.mediaserver_url"
-											alt=""
-										/>
-
-										<app-video
-											v-else
-											class="-video"
-											:poster="item.mediaserver_url"
-											:webm="item.mediaserver_url_webm"
-											:mp4="item.mediaserver_url_mp4"
-											show-loading
-										/>
-									</app-media-item-backdrop>
-								</app-responsive-dimensions>
-
+								<app-media-item-post :media-item="item" />
 								<br />
 							</div>
 						</div>
